v1.0.1
v1.0.1 lets operators set a fee recipient override, as well as a chance for this override to happen. This could be used as a "fee" for maintenance or to discourage long-term usage of the program.
Flags for this are `fee_override_address` and `fee_override_chance)
v1.0.1 also adds the `log-level` flag, and logs more things.